requirements.txt
streamlit>=1.28.0
google-generativeai